KELT light curves (not including KELT Praesepe data)

The NASA Exoplanet Archive provides access to KELT time series through a search interface and the application programming interface (API). The following table lists all the data columns in the KELT Time Series table (kelttimeseries) that can be returned through the API. See the KELT page for more information on the KELT data products available at the NASA Exoplanet Archive.

exo_kelt(kelt_field = NULL, progress = TRUE, col_spec = FALSE)

Arguments

kelt_field
KELT field of observation. Fields available are N02, N04, N06, N08, N10 and N12.

progress
If FALSE, suppresses progress of request (unix OS only).

col_spec
If FALSE, suppresses column specification message from readr (unix OS only).